Fix mdbx resource temporarily unavailable error
