+++ internal/db/mdbx_env.go
@@ -29,30 +29,6 @@ func InitializeMDBXEnv(path string, optMaxDB uint64, flags uint) (err error) {
 	return err
 }
 
-func NewMDBXEnv(path string, optMaxDB uint64, flags uint) (*mdbx.Env, error) {
-	env, err := mdbx.NewEnv()
-	if err != nil {
-		// notest
-		return nil, err
-	}
-	err = env.SetOption(mdbx.OptMaxDB, optMaxDB)
-	if err != nil {
-		// notest
-		return nil, err
-	}
-	err = env.SetGeometry(268435456, 268435456, 25769803776, 268435456, 268435456, 4096)
-	if err != nil {
-		// notest
-		return nil, err
-	}
-	err = env.Open(path, flags|mdbx.Exclusive, 0o664)
-	if err != nil {
-		// notest
-		return nil, err
-	}
-	return env, nil
-}
-
 // GetMDBXEnv returns the Juno LMDB environment. If the environment is not
 // initialized then ErrEnvNoInitialized is returned.
 func GetMDBXEnv() (*mdbx.Env, error) {
